Fixes #123 - for Rev
The bug was caused by objectiveless antags. I could've tinkered with adding objectivelessness checks everywhere, but I reckoned that taking down heads is the entire point of the mode and it's the hows and whys that remain up to players, as they did before, so I removed them from the mode code altogether.