Remaps Syndicate base armory  + Body bag box  + Six security belts  + Two stun batons  + Grenade launcher  + Tear gas grenade box  + Four "Knights" (new seven-round, .45 caliber silenced, lethal weapons, sprites and idea courtesy of SkyeAuroline)
The point of this update is to give nuke ops an option for a stealthy gun, if they want to play assassin for some reason. Also better crowd-control abilities with the tear gas and grenade launcher. Also to make the suit area more organized, as the softsuits (why do they even have these?) were opposite from the hardsuits, away from both the jetpacks and air tanks.