Unfucking custom food/drink code. Colors now work properly (for some reason old code always failed a check, so always picked a RGBY color randomly, not a defined filling color) and custom foods made with custom foods inherit colors from the components somewhat.
If you make a ice chili and mushroom and tomato soup soup soup it'll probably be a brownish-black mess, because it uses color multiplication. We're on the cutting edge of realism here!

Icon edits: makes gummy bears and jellybeans giant no more (shrunk the icon to 16x16 and redrew based on that) - not a terribly major quality loss in the latter's case. Made soup overlays very slightly transparent. Because soup.