Makes Vampire code compile. Porting is NOT finished yet thanks to some more noodly procs.
Also turns coffinsleep from an ugly hook in life.dm into a 'power' that is an unholy fusion of xenomorph regen code and ling fakedeath and can be woken up from at will, but it will inflict brief drowsiness. Or, indeed, stay in it however long it takes to regenerate.