changes to R&D and toxins
Map changes to R&D and Toxins. R&D is functionally the same with only
minor changes to the placement of the equipment and tables to make it
easier to not get yourself stuck and generally seems more open.

Toxins Lab has been significantly altered, especially the canister room,
but not so much that it also changes the surrounding area by much. Major
changes include, the addition of 6 gas coolers for cooling whatever gas
you want to cool, a reworked burn chamber (inspired by the one on hippie
station) where it uses heat transfer pipes to heat gasses rather then
burning the gasses directly so that you can heat a variety of gasses if
you wished to do so and a addition of a small area that connects to the
exit, canister room and toxins lab. This is a complete change in the
layout and design of the Toxins Lab without significantly altering the
areas around it or its location.

I hope these changes will help in general and also make it easier for
toxins techs (or antags) to do their jobs and experiment with the new
changes.