Increases Roundstart Gear Points (Hopefully)
Hopefully this will increase the number of points available for roundstart gear from 7 to 10. From what I heard bay refactored items to be cheaper but some people wish the limit was higher for things like tuxedos which cost 3 points for some reason. This should not change the meta as useful items, like webbing and holsters cannot be stacked anyways