Adds Kimono and Schoolgirl Uniform to Roundstart Gear (hopefully
A long wanted addition (by me) for kimonos and schoolgirl uniforms to be given at roundstart via the gear loadout feature. If you decide this is too much anime to commit that is fine but I figured I would try to make the necessary changes to add in the option anyways to learn.

If by some miracle this makes it past corai can someone doublecheck that I didn't accidentally delete system32 or something that would be great.