Restores all the shuttles we used to have; you'll need to manually initialize event shuttles like the train for now.
Rewrites Scomtime; it references the mode in a not-very-elegant way, but it's marginally better than duplicating the entire block of code twice.

Speaking of which, SCOM outfits moved into proper outfit defines, but since equipping them was and is broken (runtime, cannot execute) it's currently worked around... by deliberately stripping the characters naked and making them pick up their gear from lockers :^). This is resolved with some more remaps to Z2.