huge commit
adds a new deck, at the top. it is home to docking areas, the AI core, a ton of maint, solars, and other fun things

AI core is now on the new top deck, old AI core area now has the vault and a safe room. Old vault area is now the Counselor's office

adds the Counselor job, which is a hybrid between the chaplain and the psychiatrist. Both chaplain and psychiatrist are alt-titles for Counselor as well.

visually revamps a bunch of different areas