feature: Config-ifies a bunch of Hunger-related values. Balance tweaks.
Most Hungers give more Suppression by default now.
Made Vamp fangs actually usable for blood-drinking (aim for head).
Made regen much more powerful (up to a *full-on revive*) but more expensive and does NOT work on fire wounds (because both vamps and The Thing are weak to fire so it fits).