AUF Ranks
Adds the AUF Branch to the faction options, with three ranks for your adminbus or antaggery pleasure. Also adds Enlisted and Officer pins to the defines of the ICCGN and SCGA ranks, for consistency with the fleet ones.

Undoes the error, again

Oops. I hope I can squash it without undoing the commit this time

Spook fix