RIG verbs
Adds toggle verbs to take off RIG parts when the suit is offline.

Currently seems like an oversight, as the suits can do this and have been able to do this for over a year but only through unequipping the module.

Also makes RIG interactions SKILL_EVA dependent. Functions as fast as before if you have Trained EVA, and is faster/slower for higher/lower skill levels.