feature: Suppresses noncritical GOAI WIP warns unless a compile option is toggled.
CONTEXT: I've gotten in the habit of using warn directives as nice markers of things to finish/clean up as the compiler will pester me to handle them and give me a link directly to code. The warns covered by this change are exclusively WIP stuff that doesn't break anything. Keeps CI happy but provides an easy way to reenable them.